change getRemote() to match dg rig.
Hey there, I was using this code in our interactive rig. Was getting stuck in a do loop when the s3 bucket grew past 1000 items. Porting over the getRemote() code from the s3.js file in the dailygraphics rig fixed the problem. Basically, I don't think the program was correctly getting the next page of 1000 results.